bugfix for legacy filter item
restored the old filter item class which is using the
old filter item interface. If the old filter item
class would only inherit from the new one we wouldn't
have the old filter item interface but the new one.
That would lead to problems when using
BaseFilter:addFilterItem as this methods expects an object
with the old filter item interface class.