[BUGFIX] Use a proper class alias map for removed classes
In order to allow tools like PHPStan to not crash, deprecated classes
that have been removed but should still be usable now are properly
mapped using a class alias map (for production use) and a second
file that provides those classes for IDEs.

This is the recommended way to do this for TYPO3 projects.